Create redux.js
The core concepts of Redux. 

In this project  was building an adventure game using reducers, state, and actions. 
The state will represent the state of the game.
It contains the player’s inventory, distance travelled, and time on the road. 
Each event in the game is represented as an action. 
Players can gather supplies, travel, and–if they play risky–sometimes tip over the wagon carrying their supplies.